+++ types/metasql.d.ts
@@ -1,1 +1,3 @@
-export class Database {}
+export class Database {
+  col(table: string, condition?: object): Promise<number>;
+}