+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased][unreleased]
 
+- New API methods for Database:
+  - async count(table, condition): number
+
 ## [1.1.4][] - 2021-06-30
 
 - Support `OR` in queries